% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alaska}
\alias{alaska}
\title{Absorbance Dataset From Alaska}
\format{
A data frame with 193846 rows and 4 columns.
\describe{
\item{sample_id}{Core Sample ID referring to location and content of sample}
\item{bsi}{Percentage of Biogenic Silica in Sample, determined by a wet chemical process}
\item{wavenumber}{Wavenumber value (units = reciprocal centimeters)}
\item{absorbance}{Absorbance levels}
}
}
\source{
Gathered from \url{https://github.com/sds-capstone/PLSmodel}
}
\usage{
alaska
}
\description{
A dataset containing absorbance spectra for 28 Greenland lake sediment core samples across 3697 wavenumbers.
}
\keyword{datasets}
