% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftirs.R
\name{arctic_mod}
\alias{arctic_mod}
\title{A function that returns the PLSR model used by \code{predict.ftirs()} to predict BSi percentages from testing data.
This model is trained on arctic lake core samples from Alaska and Greenland.}
\usage{
arctic_mod(...)
}
\arguments{
\item{...}{Other arguments passed on to methods. Not currently used.}
}
\description{
A function that returns the PLSR model used by \code{predict.ftirs()} to predict BSi percentages from testing data.
This model is trained on arctic lake core samples from Alaska and Greenland.
}
