% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftirs.R
<<<<<<< HEAD:man/as.ftirs.Rd
\name{as.ftirs}
\alias{as.ftirs}
\title{Method that coerces data frame into object class \code{ftirs}
=======
\name{as_ftirs}
\alias{as_ftirs}
\title{Coerce data frame into object class \code{ftirs}
>>>>>>> db26ef2ca6653b0870fbc8453a16377bb30315ad:man/as_ftirs.Rd
This only changes the class label of the object in order to access the methods of the class. It does not change anything about the object besides the classification.}
\usage{
as.ftirs(df)
}
\arguments{
\item{df}{A data.frame to coerce to class \code{ftirs}.}
}
\description{
<<<<<<< HEAD:man/as.ftirs.Rd
Method that coerces data frame into object class \code{ftirs}
=======
Coerce data frame into object class \code{ftirs}
>>>>>>> db26ef2ca6653b0870fbc8453a16377bb30315ad:man/as_ftirs.Rd
This only changes the class label of the object in order to access the methods of the class. It does not change anything about the object besides the classification.
}
