% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftirs.R
<<<<<<< HEAD:man/as.ftirs.Rd
\name{as.ftirs}
\alias{as.ftirs}
=======
\name{as_ftirs}
\alias{as_ftirs}
>>>>>>> 6317200657b7e3388e072696a058346b3859bf22:man/as_ftirs.Rd
\title{Coerce data frame into object class \code{ftirs}
This only changes the class label of the object in order to access the methods of the class. It does not change anything about the object besides the classification.}
\usage{
as_ftirs(df)
}
\arguments{
\item{df}{A data.frame to coerce to class \code{ftirs}.}
}
\description{
Coerce data frame into object class \code{ftirs}
This only changes the class label of the object in order to access the methods of the class. It does not change anything about the object besides the classification.
}
