% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftirs.R
\name{ftirs}
\alias{ftirs}
\title{FTIRS class
A class attribute that is a modified dataframe that is in the proper format to be used in a Partial Least Squares Regression model and has access to the relevant class methods.}
\description{
FTIRS class
A class attribute that is a modified dataframe that is in the proper format to be used in a Partial Least Squares Regression model and has access to the relevant class methods.
}
