% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_df.R
\name{generate_df}
\alias{generate_df}
\title{A function that generates a dataframe in the proper format for the plsr model
note to specify how to format data}
\usage{
generate_df(
  your_csv_folder,
  wavenumbers = "AS-01 (8_24_16).0.csv",
  wet_chem_csv = NULL
)
}
\arguments{
\item{your_csv_folder}{The filepath to the your folder that contains the csv files for all your FTIR spectroscopy samples}

\item{wavenumbers}{The filepath to the sample with wavenumbers you want to use}
}
\description{
A function that generates a dataframe in the proper format for the plsr model
note to specify how to format data
}
