% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{greenland}
\alias{greenland}
\title{Absorbance Sample From Greenland}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 28 rows and 3700 columns.
}
\usage{
greenland
}
\description{
Absorbance Sample From Greenland
}
\keyword{datasets}
