% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation.R
\name{interpolate}
\alias{interpolate}
\title{Interpolates a dataset to have uniform wavenumbers}
\usage{
interpolate(wavenumber, absorbance, out_vec)
}
\description{
Interpolates a dataset to have uniform wavenumbers
}
