% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{interpolate_ftirs}
\alias{interpolate_ftirs}
\title{A function to interpolate a dataset}
\usage{
interpolate_ftirs(
  wavenumber,
  absorbance,
  out_vec = rounded_wavenumbers$wavenumber
)
}
\arguments{
\item{wavenumber}{The wavenumber vector}

\item{absorbance}{The absorbance vector}

\item{out_vec}{The wavenumber vector to be interpolated on}
}
\description{
A function to interpolate a dataset
}
