% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftirs.R
\name{is_ftirs}
\alias{is_ftirs}
\title{Check if an object has the FTIRS class format}
\usage{
is_ftirs(obj, ...)
}
\arguments{
\item{obj}{any R object}

\item{...}{Other arguments passed on to methods. Not currently used.}
}
\description{
Check if an object has the FTIRS class format
}
