% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\docType{data}
\name{out_vec}
\alias{out_vec}
\title{A function to interpolate a dataset}
\format{
An object of class \code{numeric} of length 1882.
}
\usage{
out_vec
}
\arguments{
\item{wavenumber}{The wavenumber vector}

\item{absorbance}{The absorbance vector}

\item{out_vec}{The wavenumber vector to be interpolated on}
}
\description{
A function to interpolate a dataset
}
\keyword{datasets}
