% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftirs.R
\name{pivot_ftirs_longer}
\alias{pivot_ftirs_longer}
\title{Function that pivots a wide ftirs dataframe back to a long format that is easier to store.}
\usage{
pivot_ftirs_longer(ftirs_data_wide, wet_chem, ...)
}
\arguments{
\item{ftirs_data_wide}{A wide ftirs dataframe that has a column for each wavenumber and a row for each sample.}

\item{wet_chem}{A logical value that states if there is a column including wet chemistry data in the wide ftirs dataframe the user is inputting. \code{TRUE} means there is such column, \code{FALSE} denotes there is not.}
}
\description{
Function that pivots a wide ftirs dataframe back to a long format that is easier to store.
}
