% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftirs.R
\name{pivot_ftirs_wider}
\alias{pivot_ftirs_wider}
\title{Function that pivots the ftirs df to wider format that is necessary for the PLS model}
\usage{
pivot_ftirs_wider(ftirs_data_long, ...)
}
\arguments{
\item{ftirs_data_long}{A long ftirs dataframe that contains columns for wavenumber and absorbance.}
}
\description{
Function that pivots the ftirs df to wider format that is necessary for the PLS model
}
