% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftirs.R
\name{pivot_longer.ftirs}
\alias{pivot_longer.ftirs}
\title{Pivot a wide, non-tidy FTIRS dataframe to a long, tidy format.}
\usage{
\method{pivot_longer}{ftirs}(ftirs_data_wide, wet_chem, ...)
}
\arguments{
\item{ftirs_data_wide}{A wide, non-tidy FTIRS dataframe. Columns = wavenumber, rows = sample_id, and values = absorbance.}

\item{wet_chem}{A logical value (\code{TRUE} or \code{FALSE}) indicating presence of Wet Chemistry Data in the wide FTIRS dataframe.}

\item{...}{Other arguments passed on to methods. Not currently used.}
}
\description{
Pivot a wide, non-tidy FTIRS dataframe to a long, tidy format.
}
