% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftirs.R
\name{pivot_wider.ftirs}
\alias{pivot_wider.ftirs}
\title{Pivot a FTIRS dataframe to wider, non-tidy format, necessary for input into a PLSR model.}
\usage{
\method{pivot_wider}{ftirs}(ftirs_data_long, ...)
}
\arguments{
\item{ftirs_data_long}{A long, tidy format FTIRS dataframe.}

\item{...}{Other arguments passed on to methods. Not currently used.}
}
\description{
Pivot a FTIRS dataframe to wider, non-tidy format, necessary for input into a PLSR model.
}
