% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_mod.R
\name{predict.ftirs}
\alias{predict.ftirs}
\title{A function that predicts bsi content based on our model with your data}
\usage{
\method{predict}{ftirs}(your_data, ...)
}
\arguments{
\item{your_data}{must be in the wide format -> looks like it might not have to be!}
}
\description{
A function that predicts bsi content based on our model with your data
}
