% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_mod.R
\name{predict_bsi}
\alias{predict_bsi}
\title{A function that predicts bsi content based on our model with your data}
\usage{
predict_bsi(your_data, ...)
}
\arguments{
\item{your_data}{must be in the wide format}
}
\description{
A function that predicts bsi content based on our model with your data
}
