% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_mod.R
\name{predict_ftirs}
\alias{predict_ftirs}
\title{A function that predicts bsi content based on our model with your data}
\usage{
predict_ftirs(your_data, ...)
}
\arguments{
\item{your_data}{must be in the wide format -> looks like it might not have to be!}

\item{...}{Other arguments passed on to methods. Not currently used.}
}
\description{
A function that predicts bsi content based on our model with your data
}
