% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftirs.R
\name{read_wet_chem}
\alias{read_wet_chem}
\title{Read and attach Wet Chemistry data to an FTIRS object}
\usage{
read_wet_chem(filepath, data, ...)
}
\arguments{
\item{filepath}{An optional filepath to singular Wet Chemistry Data file to be included in the FTIRS dataframe. At present development, this file must be a .csv with a column containing a row for each sample and its corresponding Wet Chem BSi percentage.}

\item{data}{The corresponding FTIRS dataframe to have the Wet Chemistry Data attached to.}

\item{...}{Other arguments passed on to \code{read_csv()}.}
}
\description{
This function is called in \code{read_ftirs()} via the optional \code{wet_chem_path} argument.
}
