% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftirs.R
\name{arctic_mod}
\alias{arctic_mod}
\title{Returns the PLSR model used by \code{predict.ftirs()}}
\usage{
arctic_mod()
}
\description{
This model is trained on arctic lake core samples from Alaska and Greenland.
}
