% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftirs.R
\name{as.ftirs}
\alias{as.ftirs}
\title{Method that coerces data frame into object class \code{ftirs}
This only changes the class label of the object in order to access the methods of the class. It does not change anything about the object besides the classification.}
\usage{
as.ftirs(df)
}
\arguments{
\item{df}{A data.frame to coerce to class \code{ftirs}.}
}
\description{
Method that coerces data frame into object class \code{ftirs}
This only changes the class label of the object in order to access the methods of the class. It does not change anything about the object besides the classification.
}
