% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftirs.R
\name{ftirs}
\alias{ftirs}
\alias{read_ftirs_file}
\alias{read_ftirs}
\alias{pivot_ftirs_wider}
\alias{pivot_ftirs_longer}
\title{FTIRS class}
\usage{
read_ftirs_file(single_filepath, ...)

read_ftirs(dir_path, wet_chem_path = NULL, format = "long", ...)

pivot_ftirs_wider(ftirs_data_long, ...)

pivot_ftirs_longer(ftirs_data_wide, wet_chem, ...)
}
\arguments{
\item{single_filepath}{The filepath to the your FTIR spectroscopy sample.}

\item{dir_path}{filepath to the your folder that contains the csv files for all your FTIR spectroscopy samples}

\item{wet_chem_path}{Path to wet chemistry data the user wants to include in the ftirs dataframe. This is optional.}

\item{format}{The desired format of the ftirs dataframe, either long or wide. Long is easier to store and wide is necessary to input into the model.}

\item{ftirs_data_long}{A long ftirs dataframe that contains columns for wavenumber and absorbance.}

\item{ftirs_data_wide}{A wide ftirs dataframe that has a column for each wavenumber and a row for each sample.}

\item{wet_chem}{A logical value that states if there is a column including wet chemistry data in the wide ftirs dataframe the user is inputting. \code{TRUE} means there is such column, \code{FALSE} denotes there is not.}
}
\description{
FTIRS class

A function that generates a tibble in the proper format for the plsr model
note to specify how to format data

A function that generates a dataframe in the proper format for the plsr model
note to specify how to format data

Function that pivots the ftirs df to wider format that is necessary for the PLS model

Function that pivots a wide ftirs dataframe back to a long format that is easier to store.
}
