% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{greenland}
\alias{greenland}
\title{Absorbance Sample From Greenland}
\format{
A data frame with 28 rows and 3700 columns.
\describe{
\item{dataset}{Core Sample ID}
\item{BSiPercent}{Percentage of Biogenic Silica in Sample}
\item{V1,...,V3697}{Location along Wavenumber Spectrum}
}
}
\source{
Gathered from \url{https://github.com/sds-capstone/PLSmodel}
}
\usage{
greenland
}
\description{
Absorbance Sample From Greenland
}
\keyword{datasets}
