% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{interpolate_ftirs}
\alias{interpolate_ftirs}
\title{A function to interpolate a dataset}
\usage{
interpolate_ftirs(
  wavenumber,
  absorbance,
<<<<<<< HEAD
  out_vec = read_csv("AS-01 (8_24_16).0.csv")$wavenumber
=======
  out_vec = readr::read_csv("data-raw/AS-01 (8_24_16).0.csv")$wavenumber
>>>>>>> 1f220682402e009784bc22bd485f7ea948d9c066
)
}
\arguments{
\item{wavenumber}{The wavenumber vector}

\item{absorbance}{The absorbance vector}

\item{out_vec}{The wavenumber vector to be interpolated on}

\item{wet_chem_csv}{Parameters used to interpolate wavenumber values}
}
\description{
A function to interpolate a dataset
}
