% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftirs.R
<<<<<<< HEAD:man/is.ftirs.Rd
\name{is.ftirs}
\alias{is.ftirs}
\title{Method that checks if an object has the FTIRS class format}
=======
\name{is_ftirs}
\alias{is_ftirs}
\title{Check if an object has the FTIRS class format}
>>>>>>> db26ef2ca6653b0870fbc8453a16377bb30315ad:man/is_ftirs.Rd
\usage{
is_ftirs(obj, ...)
}
\arguments{
\item{obj}{any R object}

\item{...}{Other arguments passed on to methods. Not currently used.}
}
\description{
<<<<<<< HEAD:man/is.ftirs.Rd
Method that checks if an object has the FTIRS class format
=======
Check if an object has the FTIRS class format
>>>>>>> db26ef2ca6653b0870fbc8453a16377bb30315ad:man/is_ftirs.Rd
}
