% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftirs.R
\name{predict.ftirs}
\alias{predict.ftirs}
\title{Predict percentage of BSi in samples
\code{predict.ftirs()} outputs predicted percentage of BSi in testing samples based on a model trained on lake sediment core samples from Arctic lakes in Greenland and Alaska.}
\usage{
\method{predict}{ftirs}(object, ...)
}
\arguments{
\item{object}{A wide, non-tidy \code{ftirs} dataframe.}

\item{...}{Other arguments passed on to generic predict method.}
}
\description{
Predict percentage of BSi in samples
\code{predict.ftirs()} outputs predicted percentage of BSi in testing samples based on a model trained on lake sediment core samples from Arctic lakes in Greenland and Alaska.
}
