% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftirs.R
\name{read_ftirs}
\alias{read_ftirs}
\title{Generate a tidy data frame binding multiple FTIRS samples together}
\usage{
read_ftirs(dir_path, wet_chem_path = NULL, format = "long", ...)
}
\arguments{
\item{dir_path}{Filepath to the folder that contains the csv files with FTIRS samples. Each file should be formatted such that there are three columns; index, \code{wavenumber} (numeric), and \code{absorbance} (numeric).}

\item{wet_chem_path}{An optional filepath to singular Wet Chemistry Data file to be included in the FTIRS dataframe.}

\item{format}{The desired format of the FTIRS dataframe; \code{long} (default) or \code{wide}.}

\item{...}{Other arguments passed on to \code{map_dfr()}}
}
\description{
Generate a tidy data frame binding multiple FTIRS samples together
}
