% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftirs.R
\name{read_ftirs}
\alias{read_ftirs}
\title{A function that generates a dataframe in the proper format for the plsr model
note to specify how to format data}
\usage{
read_ftirs(dir_path, wet_chem_path = NULL, ...)
}
\arguments{
\item{The}{filepath to the sample with wavenumbers you want to use}
}
\description{
A function that generates a dataframe in the proper format for the plsr model
note to specify how to format data
}
