% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftirs.R
\name{read_ftirs_file}
\alias{read_ftirs_file}
\title{A function that generates a tibble in the proper format for the plsr model
note to specify how to format data}
\usage{
read_ftirs_file(single_filepath, ...)
}
\arguments{
\item{single_filepath}{The filepath to the your FTIR spectroscopy sample.}
}
\description{
A function that generates a tibble in the proper format for the plsr model
note to specify how to format data
}
