% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftirs.R
\name{read_ftirs_file}
\alias{read_ftirs_file}
\title{Generates a tibble from a single FTIRS sample}
\usage{
read_ftirs_file(single_filepath, interpolate = TRUE, ...)
}
\arguments{
\item{single_filepath}{The filepath to an individual FTIR spectroscopy sample.}

\item{interpolate}{A logical value choosing to interpolate absorbance values onto a set of whole number wavenumbers. \code{TRUE} is default.}

\item{...}{Other arguments passed on to \code{read_csv()}.}
}
\description{
Generates a tibble from a single FTIRS sample
}
